+++ src/main/java/com/tiki/server/auth/jwt/JwtProvider.java
@@ -18,5 +18,23 @@ import static java.util.Base64.getEncoder;
 @Component
 public class JwtProvider {
 
+    @Value("${jwt.secret}")
+    private String secretKey;
 
+    public long getUserFromJwt(String token) {
+        val claims = getBody(token);
+        return Long.parseLong(claims.get("memberId").toString());
+    }
+
+    public Claims getBody(final String token) {
+        return Jwts.parserBuilder()
+                .setSigningKey(getSigningKey())
+                .build()
+                .parseClaimsJws(token)
+                .getBody();
+    }
+    private SecretKey getSigningKey() {
+        val encodedKey = getEncoder().encodeToString(secretKey.getBytes());
+        return hmacShaKeyFor(encodedKey.getBytes());
+    }
 }
