+++ src/main/java/com/tiki/server/auth/jwt/UserAuthentication.java
@@ -1,4 +1,16 @@
 package com.tiki.server.auth.jwt;
 
-public class UserAuthentication {
+import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
+import org.springframework.security.core.GrantedAuthority;
+
+import java.util.Collection;
+
+public class UserAuthentication extends UsernamePasswordAuthenticationToken {
+    public UserAuthentication(
+            Object principal,
+            Object credentials,
+            Collection<? extends GrantedAuthority> authorities
+    ) {
+        super(principal, credentials, authorities);
+    }
 }
